move market check invariant
